Renamed generic collection package to hashset
