+++ sttp/hashset/HashSet.go
@@ -21,7 +21,7 @@
 //
 //******************************************************************************************************
 
-package collection
+package hashset
 
 type void struct{}
